src/analysis/solver/simplex.ts
import { SolverState } from '../../../spec/Solver';

interface SimplexTableau {
  objective: number[];
  constraints: number[][];
  basis: number[];
  rhs: number[];
}

export function simplexMethod(
  state: SolverState,
  evaluateFunction: (variables: number[]) => number
): SolverState {
  // Convert problem to standard form
  const tableau = createTableau(state);
  
  // Solve using simplex algorithm
  const solution = solveTableau(tableau);
  
  // Update state with solution
  return updateSolution(state, solution);
}

function createTableau(state: SolverState): SimplexTableau {
  // Convert problem constraints to tableau format
  // This is a placeholder implementation
  return {
    objective: [],
    constraints: [],
    basis: [],
    rhs: []
  };
}

function solveTableau(tableau: SimplexTableau): number[] {
  // Implement simplex algorithm
  // This is a placeholder implementation
  return [];
}

function updateSolution(state: SolverState, solution: number[]): SolverState {
  // Update state with solution values
  // This is a placeholder implementation
  return state;
}