src/__tests__/analysis/solver/simplex.test.ts
import { simplexMethod } from '../../../analysis/solver/simplex';
import { createSolver, addVariable, addConstraint } from '../../../analysis/solver';

describe('Simplex Method', () => {
  test('solves maximization problem', () => {
    let state = createSolver('A1', 'maximize');
    state = addVariable(state, 'B1', { min: 0 });
    state = addVariable(state, 'B2', { min: 0 });
    state = addConstraint(state, 'B1', '<=', 4);
    state = addConstraint(state, 'B2', '<=', 6);
    
    const evaluateFunction = (variables: number[]) => {
      const [x, y] = variables;
      return 3 * x + 2 * y;
    };

    state = simplexMethod(state, evaluateFunction);
    
    // Since this is a placeholder, just verify state is returned
    expect(state).toBeDefined();
  });

  test('handles infeasible problem', () => {
    let state = createSolver('A1', 'maximize');
    state = addVariable(state, 'B1', { min: 0 });
    state = addConstraint(state, 'B1', '<=', 0);
    state = addConstraint(state, 'B1', '>=', 1);
    
    const evaluateFunction = (variables: number[]) => variables[0];

    state = simplexMethod(state, evaluateFunction);
    
    // Verify state is returned
    expect(state).toBeDefined();
  });
});